backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.4
scikit-learn==1.4.0
pandas==2.2.0
gunicorn==21.2.0